addAvatarVers1Fix#2

+++ src/main/java/ru/hogwarts/school/controller/StudentController.java
@@ -22,7 +22,7 @@ public class StudentController {
     }
 
     @GetMapping("{id}")
-    public Student getStudent(@PathVariable int id){
+    public Student getStudent(@PathVariable long id){
         return studentService.findStudent(id);
     }
 
@@ -39,7 +39,7 @@ public class StudentController {
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity deleteStudent(@PathVariable int id) {
+    public ResponseEntity deleteStudent(@PathVariable long id) {
          studentService.deleteStudent(id);
          return ResponseEntity.ok().build();
     }
